🔵refactor: CardType enum 1개를 CardScore, CardType으로 분리 및 Card 리팩토링
