+++ src/test/java/domain/CardTest.java
@@ -3,6 +3,7 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.card.Card;
+import domain.card.CardScore;
 import domain.card.CardType;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -12,10 +13,10 @@ class CardTest {
     class cardConstructor {
         @Test
         void 카드를_생성한다() {
-            Card card = new Card(CardType.CLOVER_ACE);
+            Card card = new Card(CardType.CLOVER, CardScore.ACE);
 
             assertThat(card).isInstanceOf(Card.class);
-            assertThat(card.getType()).isEqualTo(CardType.CLOVER_ACE);
+            assertThat(card.getType()).isEqualTo(CardType.CLOVER);
         }
 
         @Test
@@ -25,7 +26,7 @@ class CardTest {
             final int score = 20;
 
             //when
-            final Card card = new Card(CardType.CLOVER_ACE);
+            final Card card = new Card(CardType.CLOVER, CardScore.ACE);
             final int aceScore = card.calculateAceScore(score, limit);
 
             //then
@@ -39,7 +40,7 @@ class CardTest {
             final int score = 10;
 
             //when
-            final Card card = new Card(CardType.CLOVER_ACE);
+            final Card card = new Card(CardType.CLOVER, CardScore.ACE);
             final int aceScore = card.calculateAceScore(score, limit);
 
             //then
